+++ toolkit/offchain/src/scripts_data.rs
@@ -1,18 +1,19 @@
 use crate::{
-	csl::{plutus_script_address, plutus_script_hash},
+	csl::{ogmios_network_to_csl, plutus_script_address, plutus_script_hash},
 	untyped_plutus::{apply_params_to_script, csl_plutus_data_to_uplc, datum_to_uplc_plutus_data},
+	OffchainError,
 };
 use anyhow::anyhow;
 use cardano_serialization_lib::{LanguageKind, NetworkIdKind, PlutusData as CSLPlutusData};
 use chain_params::SidechainParams;
-use plutus::ToDatum;
+use ogmios_client::query_network::QueryNetwork;
 use serde::Serialize;
 use sidechain_domain::{MainchainAddressHash, PolicyId};
 use uplc::PlutusData;
 
 /// Provides convenient access to the addresses and hashes of the partner chain smart contracts.
 /// Data in this struct is derived from the smart contracts, applied parameters and the network.
-#[derive(Debug, Serialize, PartialEq, Eq)]
+#[derive(Clone, Debug, Serialize, PartialEq, Eq)]
 #[serde(rename_all = "camelCase")]
 pub struct ScriptsData {
 	pub addresses: Addresses,
@@ -21,7 +22,7 @@ pub struct ScriptsData {
 }
 
 /// Bech32 address of applied validators in partner-chains smart contracts.
-#[derive(Debug, Serialize, PartialEq, Eq)]
+#[derive(Clone, Debug, Default, Serialize, PartialEq, Eq)]
 #[serde(rename_all = "PascalCase")]
 pub struct Addresses {
 	pub committee_candidate_validator: String,
@@ -33,7 +34,7 @@ pub struct Addresses {
 }
 
 /// Hashes of applied validators in partner-chains smart contracts.
-#[derive(Debug, Serialize, PartialEq, Eq)]
+#[derive(Clone, Debug, Default, Serialize, PartialEq, Eq)]
 pub struct ValidatorHashes {
 	pub committee_candidate_validator: MainchainAddressHash,
 	pub d_parameter_validator: MainchainAddressHash,
@@ -44,7 +45,7 @@ pub struct ValidatorHashes {
 }
 
 /// Policy IDs of applied scripts in partner-chains smart contracts.
-#[derive(Debug, Serialize, PartialEq, Eq)]
+#[derive(Clone, Debug, Default, Serialize, PartialEq, Eq)]
 pub struct PolicyIds {
 	pub d_parameter: PolicyId,
 	pub init_token: PolicyId,
@@ -56,12 +57,36 @@ pub struct PolicyIds {
 
 type ScriptBytes = Vec<u8>;
 
-/// For the given `pc_params` it returns the addresses, hashes and policy ids of the partner chain smart contracts.
-pub fn get_scripts_data(
+pub trait GetScriptsData {
+	#[allow(async_fn_in_trait)]
+	/// For the given `pc_params` it returns the addresses, hashes and policy ids of the partner chain smart contracts.
+	async fn get_scripts_data(
+		&self,
+		pc_params: SidechainParams,
+	) -> Result<ScriptsData, OffchainError>;
+}
+
+impl<T: QueryNetwork> GetScriptsData for T {
+	async fn get_scripts_data(
+		&self,
+		pc_params: SidechainParams,
+	) -> Result<ScriptsData, OffchainError> {
+		let network = ogmios_network_to_csl(
+			self.shelley_genesis_configuration()
+				.await
+				.map_err(|e| OffchainError::OgmiosError(e.to_string()))?
+				.network,
+		);
+		get_scripts_data(pc_params, network)
+			.map_err(|e| OffchainError::InternalError(e.to_string()))
+	}
+}
+
+fn get_scripts_data(
 	pc_params: SidechainParams,
 	network: NetworkIdKind,
 ) -> anyhow::Result<ScriptsData> {
-	let pc_params_data = datum_to_uplc_plutus_data(&pc_params.to_datum());
+	let pc_params_data = datum_to_uplc_plutus_data(&pc_params);
 
 	let (version_oracle_validator, version_oracle_policy, version_oracle_policy_data) =
 		version_oracle(&pc_params_data, network)?;