+++ changelog.md
@@ -24,6 +24,7 @@ trait, implemented by separate types in `db-sync-follower` and `main-chain-follo
 on how to create these new data sources see `node/src/main_chain_follower.rs` file.
 * Added `pallet-session` integration for `pallet-session-validator-management`. Not wired in the node.
 * `partner-chains-cli` does not use `cardano-cli` to derive address not to query utxos.
+* `partner-chains-cli` does not use `pc-contracts-cli` in `prepare-configuration` wizard, it uses `partner-chains-cardano-offchain` crate instead.
 
 ## Removed